mcp_server/requirements.txt
fastapi>=0.110.0
uvicorn>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai-agents>=0.0.7
mem0ai>=1.0.0
qdrant-client>=1.17.0
